requirements.txt
flask
sqlalchemy
requests
pdfplumber==0.11.6
python-dotenv
gTTS
pytesseract
Pillow
beautifulsoup4
gunicorn
cryptography==41.0.7
pysqlite3
chromadb==0.4.24
numpy<2.0.0
SpeechRecognition
librosa
soundfile
google-cloud-translate
